% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{example_list_loc}
\alias{example_list_loc}
\title{Location of the example list}
\usage{
example_list_loc()
}
\value{
The location of the installed dataset
}
\description{
This package installs an example list of names. This command locates it in
the installed version.
}
