% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{get_one}
\alias{get_one}
\title{Get one}
\usage{
get_one(names_location = "data/list.txt")
}
\arguments{
\item{names_location}{Location of the names file}
}
\value{
One random name from the names file
}
\description{
Get one
}
