% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{make_groups}
\alias{make_groups}
\title{Make groups}
\usage{
make_groups(names_location = "data/list.txt", group_size = 2, padding = "")
}
\arguments{
\item{names_location}{Location of the names file}

\item{group_size}{Desired size for each group}

\item{padding}{Character for empty slots}
}
\value{
The names, grouped
}
\description{
Make groups
}
