% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{read_people}
\alias{read_people}
\title{Read names list}
\usage{
read_people(names_location = "data/list.txt", remove.empty = TRUE)
}
\arguments{
\item{names_location}{Location of the names file}

\item{remove.empty}{(default=TRUE) for ignoring empty lines in file}
}
\value{
A vector containing all the names
}
\description{
Read names list
}
