% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auxs.R
\name{read_people}
\alias{read_people}
\title{Read names list}
\usage{
read_people(
  names_location = example_list_loc(),
  rm.empty = TRUE,
  rm.duplicates = TRUE
)
}
\arguments{
\item{names_location}{Location of the names file}

\item{rm.empty}{(default = TRUE) for ignoring empty lines in file}

\item{rm.duplicates}{(default = TRUE) for removing duplicate entries}
}
\value{
A vector containing all the names
}
\description{
Read names list
}
